requirements.txt
requests
beautifulsoup4
PyMuPDF
Pillow
Jinja2
sendgrid
boto3
python-dotenv
PyYAML
pdf2image
playwright
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
